Updated CSS selectors for scratch blocks
Using the data-category attribute to determine the category of a block deemed unreliable (it sometimes just reports 'null'). To combat this, it now makes use of the background-color property in the style attribute of the scratch block div. Indeed not the most efficient and logical method, but the only one I found that didn't break every two minutes.